+++ torchvision/csrc/cuda/vision_cuda.h
@@ -90,12 +90,12 @@ at::Tensor DCN_forward_cuda(
     const at::Tensor& input,
     const at::Tensor& offset,
     const at::Tensor& weights,
-    const std::pair<int, int>& stride,
-    const std::pair<int, int>& pad,
-    const std::pair<int, int>& dilation,
-    const int groups,
-    const int deformable_groups,
-    const int im2col_step);
+    const std::pair<int64_t, int64_t>& stride,
+    const std::pair<int64_t, int64_t>& pad,
+    const std::pair<int64_t, int64_t>& dilation,
+    const int64_t groups,
+    const int64_t deformable_groups,
+    const int64_t im2col_step);
 
 at::Tensor DCN_backward_cuda(
     const at::Tensor& grad,
